Safari-611.1.10/LayoutTests/fast/text-autosizing/ios/idempotentmode/idempotent-autosizing-identity-expected.txt

PASS result is 12
PASS result is 6
PASS successfullyParsed is true

TEST COMPLETE
Test
Test
Test
Test
